Rename events package to listener
